+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## Current release:
-  * [CHANGELOG-1.0.md][10]
+  * [CHANGELOG-1.1.md][11]
 
 ## Development release:
   * [Unreleased Changes][0]
 
 ## Older releases:
+  * [CHANGELOG-1.0.md][10]
   * [CHANGELOG-0.11.md][9]
   * [CHANGELOG-0.10.md][8]
   * [CHANGELOG-0.9.md][7]
@@ -16,6 +17,7 @@
   * [CHANGELOG-0.3.md][1]
 
 
+[11]: https://github.com/heptio/velero/blob/master/changelogs/CHANGELOG-1.1.md
 [10]: https://github.com/heptio/velero/blob/master/changelogs/CHANGELOG-1.0.md
 [9]: https://github.com/heptio/velero/blob/master/changelogs/CHANGELOG-0.11.md
 [8]: https://github.com/heptio/velero/blob/master/changelogs/CHANGELOG-0.10.md